[cheriabi] Preserve capability provenance for enum_type.
With `G_ENABLE_EXPERIMENTAL_ABI_COMPILATION` enabled (for example, as
when compiling for aarch64c) glib no longer assumes that `size_t` and
`uintptr_t` (and therefore `GType` and `gsize`) are the same width.  In
order to preserve capability provenance the type of `g_define_type` must
match that of `enum_type`, that is `GType`.